# Remove unnecessary boxing allocation