Profiling over around 10s in `TestSceneBufferingPerformance`:

| Before | After |
| --- | --- |
| <img width="832" alt="image" src="https://github.com/user-attachments/assets/1a4931fc-f05d-4281-bc36-d03abf40aff4"> | <img width="834" alt="image" src="https://github.com/user-attachments/assets/362a189b-9015-4c82-873e-d897f658c793"> |

This is a bit of an edge case, but it is a potential source of per-frame allocations nonetheless.